added comments, tidied commented out code and filenames
