+++ index.js
@@ -1,70 +1,14 @@
-// const SVGLogos = require("./lib/logo.js");
-
-// console.log(1, 2, 3);
-
-// const logo = new SVGLogos();
-
-// logo.main();
-
-// const inquirer = require("inquirer");
 const { writeFile } = require("fs/promises");
-const getData = require("./lib/getData.js");
-// const generateShape = require("./lib/generateSVG.js");
-const Circle = require("./lib/circle.js");
-const Square = require("./lib/square.js");
-const Triangle = require("./lib/triangle.js");
-
-// class SVGLogos {
-//   constructor() {
-//     this.shape = "";
-//   }
-// }
-// function generateShape(chars, color, bgColor, structure) {
-//   switch (structure) {
-//     case "Circle":
-//       console.log("all the things", chars, color, bgColor);
-//       const newShape = new Circle(chars, color, bgColor);
-//       // console.log("x", "y", "z", newShape);
-//       return newShape;
-//     case "Square":
-//       console.log(2);
-//       break;
-//     case "Triangle":
-//       console.log(3);
-//       break;
-
-//     default:
-//       break;
-//   }
-// }
-
-const generateShape = (chars, color, bgColor, structure) => {
-  switch (structure) {
-    case "Circle":
-      const newCircle = new Circle(chars, color, bgColor);
-      return newCircle;
-    case "Square":
-      const newSquare = new Square(chars, color, bgColor);
-      return newSquare;
-    case "Triangle":
-      const newTriangle = new Triangle(chars, color, bgColor);
-      return newTriangle;
-
-    default:
-      break;
-  }
-};
+const getData = require("./lib/returnData.js");
+const generateShape = require("./lib/returnSVG.js");
 
 const main = async () => {
   try {
     const { chars, color, bgColor, structure } = await getData;
 
-    // console.log("a", "b", chars, color, bgColor, structure);
     const shape = await generateShape(chars, color, bgColor, structure);
-    // console.log("finally?", shape);
 
-    const fileContents = await shape.render();
-    // console.log(typeof fileContents);
+    const fileContents = shape.render();
 
     await writeFile("./examples/logo.svg", fileContents);
     console.log("Generated logo.svg");
@@ -104,22 +48,3 @@ main();
 // uses data to generate shape
 // generate shape runs if to select structure and thens runs relevant class constructor
 // then call render on generated shape
-
-// const generateShape = (chars, color, bgColor, structure) => {
-//   switch (structure) {
-//     case "Circle":
-//       console.log("all the things", chars, color, bgColor);
-//       const newShape = new Circle(chars, color, bgColor);
-//       console.log("x", "y", "z", newShape);
-//       return newShape;
-//     case "Square":
-//       console.log(2);
-//       break;
-//     case "Triangle":
-//       console.log(3);
-//       break;
-
-//     default:
-//       break;
-//   }
-// };
