+++ lib/classCircle.js
@@ -1,4 +1,4 @@
-const Shape = require("./shape");
+const Shape = require("./classShape-Parent");
 
 // builds on shape by giving it a structure and method to return it.
 class Circle extends Shape {
@@ -10,12 +10,12 @@ class Circle extends Shape {
     this.structure = `<svg version="1.1"
         width="300" height="200"
         xmlns="http://www.w3.org/2000/svg">
-   
-     <circle cx="150" cy="100" r="80" fill="${this.bgColor}" />
-   
-     <text x="150" y="115" font-size="60" text-anchor="middle" fill="${this.color}">${this.chars}</text>
-   
-   </svg>`;
+
+<circle cx="150" cy="100" r="80" fill="${this.bgColor}" />
+
+<text x="150" y="115" font-size="60" text-anchor="middle" fill="${this.color}">${this.chars}</text>
+
+</svg>`;
   }
 }
