+++ lib/classShape-Parent.js
@@ -10,6 +10,7 @@
 //   '<polygon points="150, 18 244, 182 56, 182" fill="blue" />'
 // );
 
+// parent class for other shapes
 class Shape {
   constructor(chars, color, bgColor) {
     this.chars = chars;
