+++ lib/classSquare.js
@@ -1,7 +1,7 @@
-const Shape = require("./shape");
+const Shape = require("./classShape-Parent");
 
 // builds on shape by giving it a structure and method to return it.
-class Triangle extends Shape {
+class Square extends Shape {
   constructor(chars, color, bgColor) {
     super(chars, color, bgColor);
 
@@ -11,13 +11,14 @@ class Triangle extends Shape {
     width="300" height="200"
     xmlns="http://www.w3.org/2000/svg">
 
-<polygon points="150 20, 50 180, 250 180" fill="${this.bgColor}" />
+<rect x="70" y="20" width="160" height="160" fill="${this.bgColor}" />
+  "/>
 
- <text x="150" y="178" font-size="60" text-anchor="middle" fill="${this.color}">${this.chars}</text>
-  </text>
+<text x="150" y="115" font-size="60" text-anchor="middle" fill="${this.color}">${this.chars}</text>
 
-</svg>`;
+</svg>
+`;
   }
 }
 
-module.exports = Triangle;
+module.exports = Square;
